requirements.txt
requests
pandas
python-dotenv
plotly
dash
pyarrow
psycopg[binary]
